# fix node issue 49960